project task 6

+++ 0x00-hello_world/6-size.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+
 /**
  * Main entry point
  *
@@ -6,17 +7,18 @@
  */
 int main(void)
 {
+	/* variable initilation*/
 	int intType;
 	float floatType;
-	double llType;
+	int llType;
 	char charType;
 	int longType;
 
-       	printf("Size of a char: %zu bytes\n", sizeof(charType));
+	printf("Size of a char: %zu bytes\n", sizeof(charType));
 	printf("Size of a int: %zu bytes\n", sizeof(intType));
 	printf("Size of a long int: %zu bytes\n", sizeof(longType));
 	printf("Size of a long long int: %zu bytes\n", sizeof(llType));
 	printf("Size of a float: %zu bytes\n", sizeof(floatType));
-	                      
-	return(0);
+
+	return (0);
 }
